Refactor redis into a separate class
